+++ gui/Aqua/SPFriendsController.m
@@ -64,7 +64,7 @@ static SPFriendsController *sharedFriendsController = nil;
     return self;
 }
 
-- (unsigned)retainCount
+- (NSUInteger)retainCount
 {
     return UINT_MAX; // denotes an object that cannot be released
 }
